Creating MicroObjects Common Library Project
Pulling in some git submodules. Fluent Types and Interface Mocks.
This should serve as an easily includable, or submodule-able, project which can be treated as "your code". The idea is to modify the project to make it exactly what is needed. When that need is recognized as becoming common, it can be PR'd into the master and that will spread to the rest of the users.